+++ src/icalendar/cal/alarm.py
@@ -259,7 +259,7 @@ class Alarm(Component):
             :class:`Alarm`
 
         Raises:
-            IncompleteComponent: If the content is not valid according to :rfc:`5545`.
+            InvalidCalendar: If the content is not valid according to :rfc:`5545`.
 
         .. warning:: As time progresses, we will be stricter with the validation.
         """