+++ src/icalendar/tests/attr/test_component.py
@@ -63,7 +63,7 @@ def test_set_value_and_get_it(
     assert in_utc == dtstamp_comp.DTSTAMP
 
 
-@pytest.mark.parametrize("invalid_value", [None, timedelta()])
+@pytest.mark.parametrize("invalid_value", ["asdasd", timedelta()])
 def test_set_invalid_value(invalid_value, dtstamp_comp):
     """Check handling of invalid values."""
     with pytest.raises(TypeError) as e:
@@ -71,7 +71,7 @@ def test_set_invalid_value(invalid_value, dtstamp_comp):
     assert e.value.args[0] == f"DTSTAMP takes a datetime in UTC, not {invalid_value}"
 
 
-@pytest.mark.parametrize("invalid_value", [None, vDDDTypes(timedelta())])
+@pytest.mark.parametrize("invalid_value", ["ashdkjasjkdkd", vDDDTypes(timedelta())])
 def test_get_invalid_value(invalid_value, dtstamp_comp):
     """Check handling of invalid values."""
     dtstamp_comp["DTSTAMP"] = invalid_value
@@ -94,3 +94,47 @@ def test_last_modified(dtstamp_comp, tzp):
     """Check we can set LAST_MODIFIED in the same way as DTSTAMP"""
     dtstamp_comp.LAST_MODIFIED = date(2014, 1, 2)
     assert tzp.localize_utc(datetime(2014, 1, 2)) == dtstamp_comp.LAST_MODIFIED
+
+
+@pytest.fixture(params=["last_modified", "created"])
+def attr_depending_on_dtstamp(request):
+    """These parameters default to the DTSTAMP attribute."""
+    return request.param
+
+
+@pytest.mark.parametrize(
+    "dt", [datetime(2011, 10, 5, 13, 32, 25), datetime(2011, 10, 5)]
+)
+def test_last_modified_defaults_to_dtstamp(dt, tzp, c, attr_depending_on_dtstamp):
+    """In the case of an iCalendar object that doesn't specify a "METHOD"
+    property, "DTSTAMP" property is equivalent to the "LAST-MODIFIED"
+    property."""
+    c.stamp = dt
+    value = getattr(c, attr_depending_on_dtstamp)
+    assert value == c.stamp
+
+
+def test_set_last_modified(c, attr_depending_on_dtstamp):
+    """Setting last modified does not override the dtstamp."""
+    setattr(c, attr_depending_on_dtstamp, datetime(2011, 10, 5, 13, 32, 25))
+    assert c.DTSTAMP is None
+
+
+def test_set_last_modified_2(c, tzp, attr_depending_on_dtstamp):
+    """Setting last modified does not override the dtstamp."""
+    c.DTSTAMP = datetime(2011, 10, 5)
+    setattr(c, attr_depending_on_dtstamp, datetime(2011, 10, 5, 13, 32, 25))
+    assert c.DTSTAMP == tzp.localize_utc(datetime(2011, 10, 5))
+
+
+def test_delete_last_modified_1(c, attr_depending_on_dtstamp):
+    """Delete the absent value."""
+    delattr(c, attr_depending_on_dtstamp)
+    assert getattr(c, attr_depending_on_dtstamp) is None
+
+
+def test_delete_last_modified_2(c, attr_depending_on_dtstamp):
+    """Delete the present value."""
+    setattr(c, attr_depending_on_dtstamp, datetime(2011, 10, 5, 13, 32, 25))
+    delattr(c, attr_depending_on_dtstamp)
+    assert getattr(c, attr_depending_on_dtstamp) is None