+++ src/icalendar/tests/conftest.py
@@ -22,6 +22,7 @@ from icalendar import (
 from icalendar.compatibility import ZoneInfo, zoneinfo
 from icalendar.timezone import TZP
 from icalendar.timezone import tzp as _tzp
+from icalendar import TypesFactory
 
 from . import timezone_ids
 
@@ -105,6 +106,7 @@ CALENDARS_FOLDER = HERE / "calendars"
 TIMEZONES_FOLDER = HERE / "timezones"
 EVENTS_FOLDER = HERE / "events"
 ALARMS_FOLDER = HERE / "alarms"
+AVAILABILITIES_FOLDER = HERE / "availabilities"
 
 
 @pytest.fixture(scope="module")
@@ -125,6 +127,11 @@ def events(tzp):
 @pytest.fixture(scope="module")
 def alarms(tzp):
     return DataSource(ALARMS_FOLDER, Alarm.from_ical)
+
+
+@pytest.fixture(scope="module")
+def availabilities(tzp):
+    return DataSource(AVAILABILITIES_FOLDER, icalendar.cal.alarm.Alarm.from_ical)
 
 
 @pytest.fixture(params=PYTZ_UTC + [ZoneInfo("UTC"), tz.UTC, tz.gettz("UTC")])
@@ -187,6 +194,7 @@ def types_factory():
     """Return a new types factory."""
     return TypesFactory()
 
+
 @pytest.fixture
 def x_sometime(types_factory):
     """Map x_sometime to time"""
@@ -231,6 +239,15 @@ def calendar_component(tzp):
     c = Component()
     c.name = "VCALENDAR"
     return c
+
+
+@pytest.fixture
+def dont_validate_new():
+    """Remove validation for new() of components."""
+    value = Component._validate_new
+    Component._validate_new = False
+    yield
+    Component._validate_new = value
 
 
 @pytest.fixture
@@ -348,6 +365,12 @@ def doctest_import(name, *args, **kw):
 NOW = datetime(2025, 5, 17, 8, 6, 12)
 NOW_UTC = NOW.replace(tzinfo=timezone.utc)
 UID_DEFAULT = "d755cef5-2311-46ed-a0e1-6733c9e15c63"
+
+
+@pytest.fixture
+def test_uid() -> str:
+    """The UID that tests always create."""
+    return UID_DEFAULT
 
 
 @pytest.fixture(autouse=True)